Extract common claim provider methods to mixin
These are shared between all advocate/agfs claims
types and the advent of a new hardship claim
has prompted this refactor.
