+++ app/models/claim/advocate_claim.rb
@@ -1,71 +1,8 @@
-# == Schema Information
-#
-# Table name: claims
-#
-#  id                       :integer          not null, primary key
-#  additional_information   :text
-#  apply_vat                :boolean
-#  state                    :string
-#  last_submitted_at        :datetime
-#  case_number              :string
-#  advocate_category        :string
-#  first_day_of_trial       :date
-#  estimated_trial_length   :integer          default(0)
-#  actual_trial_length      :integer          default(0)
-#  fees_total               :decimal(, )      default(0.0)
-#  expenses_total           :decimal(, )      default(0.0)
-#  total                    :decimal(, )      default(0.0)
-#  external_user_id         :integer
-#  court_id                 :integer
-#  offence_id               :integer
-#  created_at               :datetime
-#  updated_at               :datetime
-#  valid_until              :datetime
-#  cms_number               :string
-#  authorised_at            :datetime
-#  creator_id               :integer
-#  evidence_notes           :text
-#  evidence_checklist_ids   :string
-#  trial_concluded_at       :date
-#  trial_fixed_notice_at    :date
-#  trial_fixed_at           :date
-#  trial_cracked_at         :date
-#  trial_cracked_at_third   :string
-#  source                   :string
-#  vat_amount               :decimal(, )      default(0.0)
-#  uuid                     :uuid
-#  case_type_id             :integer
-#  form_id                  :string
-#  original_submission_date :datetime
-#  retrial_started_at       :date
-#  retrial_estimated_length :integer          default(0)
-#  retrial_actual_length    :integer          default(0)
-#  retrial_concluded_at     :date
-#  type                     :string
-#  disbursements_total      :decimal(, )      default(0.0)
-#  case_concluded_at        :date
-#  transfer_court_id        :integer
-#  supplier_number          :string
-#  effective_pcmh_date      :date
-#  legal_aid_transfer_date  :date
-#  allocation_type          :string
-#  transfer_case_number     :string
-#  clone_source_id          :integer
-#  last_edited_at           :datetime
-#  deleted_at               :datetime
-#  providers_ref            :string
-#  disk_evidence            :boolean          default(FALSE)
-#  fees_vat                 :decimal(, )      default(0.0)
-#  expenses_vat             :decimal(, )      default(0.0)
-#  disbursements_vat        :decimal(, )      default(0.0)
-#  value_band_id            :integer
-#  retrial_reduction        :boolean          default(FALSE)
-#  prosecution_evidence     :boolean          default(FALSE)
-#
-
 module Claim
   class AdvocateClaim < BaseClaim
     route_key_name 'advocates_claim'
+
+    include ProviderDelegation
 
     has_many :basic_fees,
              foreign_key: :claim_id,
@@ -234,34 +171,6 @@ module Claim
     end
 
     private
-
-    # TODO: SUPPLEMENTARY_CLAIM_TODO mixin/concern Claims::AdvocateClaimProviderDelegation??
-    def provider_delegator
-      if provider.firm?
-        provider
-      elsif provider.chamber?
-        external_user
-      else
-        raise "Unknown provider type: #{provider.provider_type}"
-      end
-    end
-
-    # TODO: SUPPLEMENTARY_CLAIM_TODO mixin/concern Claims::AdvocateClaimProviderDelegation??
-    def agfs_supplier_number
-      if provider.firm?
-        provider.firm_agfs_supplier_number
-      else
-        external_user.supplier_number
-      end
-    rescue StandardError
-      nil
-    end
-
-    # TODO: SUPPLEMENTARY_CLAIM_TODO mixin/concern Claims::AdvocateClaimProviderDelegation??
-    def set_supplier_number
-      supplier_no = agfs_supplier_number
-      self.supplier_number = supplier_no if supplier_number != supplier_no
-    end
 
     # create a blank fee for every basic fee type not passed to Claim::AdvocateClaim.new
     def instantiate_basic_fees
